Make dynd a namespace package.
